+++ include/dirc_babar_digitizer.h
@@ -10,25 +10,25 @@
 class DircBaBarDigitizer : public DircDigitizer
 {
 private:
-	double minx,maxx,miny,maxy;
-	double pmt_r;
-	double pmt_sep_x,pmt_sep_y;
-	double second_grid_minx,second_grid_miny;
-	double resx, resy;
-	double t_unc;
-	double t_bin_size;
+	float minx,maxx,miny,maxy;
+	float pmt_r;
+	float pmt_sep_x,pmt_sep_y;
+	float second_grid_minx,second_grid_miny;
+	float resx, resy;
+	float t_unc;
+	float t_bin_size;
 	
 	TRandom3 *dig_rand;
 	
 public:
 	DircBaBarDigitizer(\
-		double iminx,\
-		double imaxx,\
-		double iminy,\
-		double imaxy,\
-		double ipmt_r,\
-		double it_unc,\
-		double it_bin_size);
+		float iminx,\
+		float imaxx,\
+		float iminy,\
+		float imaxy,\
+		float ipmt_r,\
+		float it_unc,\
+		float it_bin_size);
 
 	void digitize_point(dirc_point &pt);
 	//return negative time if point is lost
